+++ packages/angular-workspace/example-client-app/src/app/app.module.ts
@@ -34,7 +34,9 @@ import { NimbleTableColumnMenuButtonModule } from '@ni/nimble-angular/table-colu
 import { NimbleRichTextViewerModule } from '@ni/nimble-angular/rich-text/viewer';
 import { NimbleRichTextEditorModule } from '@ni/nimble-angular/rich-text/editor';
 import { NimbleRichTextMentionUsersModule } from '@ni/nimble-angular/rich-text-mention/users';
-import { OkButtonModule } from 'ok-angular/button/ok-button.module';
+import { NimbleStepperModule } from '@ni/nimble-angular/stepper';
+import { NimbleStepModule } from '@ni/nimble-angular/step';
+import { NimbleAnchorStepModule } from '@ni/nimble-angular/anchor-step';
 import { SprightChatConversationModule } from '@ni/spright-angular/chat/conversation';
 import { SprightChatInputModule } from '@ni/spright-angular/chat/input';
 import { SprightIconWorkItemCalendarWeekDirective } from '@ni/spright-angular/icons/work-item-calendar-week';
@@ -44,11 +46,88 @@ import { SprightChatMessageSystemModule } from '@ni/spright-angular/chat/message
 import { SprightRectangleModule } from '@ni/spright-angular/rectangle';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
+import { SubContainerModule } from './customapp/sub-container/sub-container.module';
+import { OverviewSectionComponent } from './customapp/overview-section.component';
+import { AnchorSectionComponent } from './customapp/anchor-section.component';
+import { BannerSectionComponent } from './customapp/banner-section.component';
+import { BreadcrumbSectionComponent } from './customapp/breadcrumb-section.component';
+import { ButtonsSectionComponent } from './customapp/buttons-section.component';
+import { AnchorButtonsSectionComponent } from './customapp/anchor-buttons-section.component';
+import { ToggleButtonsSectionComponent } from './customapp/toggle-buttons-section.component';
+import { CardSectionComponent } from './customapp/card-section.component';
+import { CardButtonSectionComponent } from './customapp/card-button-section.component';
+import { CheckboxSectionComponent } from './customapp/checkbox-section.component';
+import { ChipSectionComponent } from './customapp/chip-section.component';
+import { RadioButtonsSectionComponent } from './customapp/radio-buttons-section.component';
+import { DialogSectionComponent } from './customapp/dialog-section.component';
+import { DrawerSectionComponent } from './customapp/drawer-section.component';
+import { MenuSectionComponent } from './customapp/menu-section.component';
+import { MenuButtonSectionComponent } from './customapp/menu-button-section.component';
+import { NumberFieldSectionComponent } from './customapp/number-field-section.component';
+import { SelectSectionComponent } from './customapp/select-section.component';
+import { ComboboxSectionComponent } from './customapp/combobox-section.component';
+import { RichTextEditorSectionComponent } from './customapp/rich-text-editor-section.component';
+import { RichTextViewerSectionComponent } from './customapp/rich-text-viewer-section.component';
+import { SpinnerSectionComponent } from './customapp/spinner-section.component';
+import { StepperSectionComponent } from './customapp/stepper-section.component';
+import { StepperAnchorSectionComponent } from './customapp/stepper-anchor-section.component';
+import { SwitchSectionComponent } from './customapp/switch-section.component';
+import { TableSectionComponent } from './customapp/table-section.component';
+import { DelayedHierarchyTableSectionComponent } from './customapp/delayed-hierarchy-table-section.component';
+import { TabsSectionComponent } from './customapp/tabs-section.component';
+import { AnchorTabsSectionComponent } from './customapp/anchor-tabs-section.component';
+import { TextAreaSectionComponent } from './customapp/text-area-section.component';
+import { TextFieldSectionComponent } from './customapp/text-field-section.component';
+import { ToolbarSectionComponent } from './customapp/toolbar-section.component';
+import { TooltipSectionComponent } from './customapp/tooltip-section.component';
+import { TreeViewSectionComponent } from './customapp/tree-view-section.component';
+import { ChatConversationSectionComponent } from './customapp/chat-conversation-section.component';
+import { IconsSectionComponent } from './customapp/icons-section.component';
+import { RectangleSectionComponent } from './customapp/rectangle-section.component';
+import { ExSectionModule } from './customapp/ex/ex-section.module';
+import { FvSectionModule } from './customapp/fv/fv-section.module';
 
 @NgModule({
     declarations: [
         AppComponent,
-        CustomAppComponent
+        CustomAppComponent,
+        OverviewSectionComponent,
+        AnchorSectionComponent,
+        BannerSectionComponent,
+        BreadcrumbSectionComponent,
+        ButtonsSectionComponent,
+        AnchorButtonsSectionComponent,
+        ToggleButtonsSectionComponent,
+        CardSectionComponent,
+        CardButtonSectionComponent,
+        CheckboxSectionComponent,
+        ChipSectionComponent,
+        RadioButtonsSectionComponent,
+        DialogSectionComponent,
+        DrawerSectionComponent,
+        MenuSectionComponent,
+        MenuButtonSectionComponent,
+        NumberFieldSectionComponent,
+        SelectSectionComponent,
+        ComboboxSectionComponent,
+        RichTextEditorSectionComponent,
+        RichTextViewerSectionComponent,
+        SpinnerSectionComponent,
+        StepperSectionComponent,
+        StepperAnchorSectionComponent,
+        SwitchSectionComponent,
+        TableSectionComponent,
+        DelayedHierarchyTableSectionComponent,
+        TabsSectionComponent,
+        AnchorTabsSectionComponent,
+        TextAreaSectionComponent,
+        TextFieldSectionComponent,
+        ToolbarSectionComponent,
+        TooltipSectionComponent,
+        TreeViewSectionComponent,
+        ChatConversationSectionComponent,
+        IconsSectionComponent,
+        RectangleSectionComponent,
     ],
     imports: [
         BrowserModule,
@@ -120,7 +199,8 @@ import { CustomAppComponent } from './customapp/customapp.component';
         NimbleMappingEmptyModule,
         NimbleIconPencilToRectangleModule,
         NimbleIconMessagesSparkleModule,
-        OkButtonModule,
+        FvSectionModule,
+        ExSectionModule,
         SprightChatConversationModule,
         SprightChatInputModule,
         SprightChatMessageInboundModule,
@@ -128,6 +208,10 @@ import { CustomAppComponent } from './customapp/customapp.component';
         SprightChatMessageSystemModule,
         SprightIconWorkItemCalendarWeekDirective,
         SprightRectangleModule,
+        NimbleStepperModule,
+        NimbleStepModule,
+        NimbleAnchorStepModule,
+        SubContainerModule,
         RouterModule.forRoot(
             [
                 { path: '', redirectTo: '/customapp', pathMatch: 'full' },