+++ packages/angular-workspace/nimble-angular/stepper/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-stepper.directive';
+export * from './nimble-stepper.module';