+++ packages/blazor-workspace/SprightBlazor/Source/Icons/SprightIconNigelChat.razor.cs
@@ -1,0 +1,5 @@
+namespace SprightBlazor;
+
+public partial class SprightIconNigelChat : SprightIconBase
+{
+}