+++ packages/nimble-components/src/label-provider/core/label-token-defaults.ts
@@ -4,8 +4,8 @@ type TokenName = keyof typeof TokensNamespace;
 
 export const coreLabelDefaults: { readonly [key in TokenName]: string } = {
     popupDismissLabel: 'Close',
-    numericIncrementLabel: 'Increment',
     numericDecrementLabel: 'Decrement',
+    numericIncrementLabel: 'Increment',
     popupIconErrorLabel: 'Error',
     popupIconWarningLabel: 'Warning',
     popupIconCompletedLabel: 'Completed',