+++ packages/react-workspace/nimble-react/src/step/index.ts
@@ -1,0 +1,6 @@
+import { Step, stepTag } from '@ni/nimble-components/dist/esm/step';
+import { wrap } from '../utilities/react-wrapper';
+
+export { stepTag };
+export { type Step };
+export const NimbleStep = wrap(Step);