+++ packages/storybook/src/docs/component-data/ok/fv/index.ts
@@ -1,0 +1,15 @@
+import { ComponentFrameworkStatus } from '../../../types';
+
+export const componentDataOkFv = [
+    {
+        componentName: 'Fv Accordion Item',
+        componentHref: './?path=/docs/ok-fv-accordion-item--docs',
+        designHref: 'https://www.figma.com/design/PO9mFOu5BCl8aJvFchEeuN/Nimble_Components?node-id=4618-64645',
+        designLabel: 'Figma',
+        library: 'ok',
+        componentStatus: ComponentFrameworkStatus.ready,
+        angularStatus: ComponentFrameworkStatus.ready,
+        blazorStatus: ComponentFrameworkStatus.ready,
+        reactStatus: ComponentFrameworkStatus.ready
+    }
+] as const;