+++ packages/blazor-workspace/SprightBlazor/CHANGELOG.md
@@ -1,8 +1,81 @@
 # Change Log - @ni/spright-blazor
 
-<!-- This log was last generated on Thu, 09 Apr 2026 19:17:36 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 23 Apr 2026 15:58:53 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 4.13.0
+
+Thu, 23 Apr 2026 15:58:53 GMT
+
+### Minor changes
+
+- Add wrapper for nigel chat icon. ([ni/nimble@9f6c8ee](https://github.com/ni/nimble/commit/9f6c8ee156880eb2cf988e34195d8c3073c1748a))
+
+## 4.12.2
+
+Wed, 22 Apr 2026 19:38:49 GMT
+
+### Patches
+
+- Blazor chat message welcome project organization ([ni/nimble@cfd38c0](https://github.com/ni/nimble/commit/cfd38c0ce1480a52731bfec197eb15b37114cbc3))
+
+## 4.12.1
+
+Tue, 21 Apr 2026 19:47:00 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.17.0
+
+## 4.12.0
+
+Mon, 20 Apr 2026 22:42:19 GMT
+
+### Minor changes
+
+- Add welcome message for spright chat conversation ([ni/nimble@f7e4581](https://github.com/ni/nimble/commit/f7e4581eb131183f6ebccd9f6337b1f54eb9be8f))
+- Bump @ni/spright-components to v6.16.0
+
+## 4.11.18
+
+Mon, 20 Apr 2026 16:30:31 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.15.9
+
+## 4.11.17
+
+Sat, 18 Apr 2026 15:44:00 GMT
+
+### Patches
+
+- Re-organize blazor file layout ([ni/nimble@08d3240](https://github.com/ni/nimble/commit/08d3240d1aac8ad1c0d9f2172ee21e8553bf9be7))
+
+## 4.11.16
+
+Wed, 15 Apr 2026 19:52:21 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.15.8
+
+## 4.11.15
+
+Wed, 15 Apr 2026 19:06:54 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.15.7
+
+## 4.11.14
+
+Tue, 14 Apr 2026 20:31:57 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.15.6
 
 ## 4.11.13