+++ packages/ok-components/CHANGELOG.md
@@ -1,8 +1,93 @@
 # Change Log - @ni/ok-components
 
-<!-- This log was last generated on Thu, 09 Apr 2026 19:17:36 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 21 Apr 2026 19:47:00 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.0.1
+
+Tue, 21 Apr 2026 19:47:00 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.17.0
+
+## 1.0.0
+
+Tue, 21 Apr 2026 14:27:32 GMT
+
+### Major changes
+
+- Rename to ok-ts-icon-dynamic. Note: dynamic icon registration prefix also updated to ok-ts-icon-dynamic- ([ni/nimble@2479cd5](https://github.com/ni/nimble/commit/2479cd5030554576e4a251c457f8a764e637cdb8))
+
+## 0.4.4
+
+Tue, 21 Apr 2026 04:58:19 GMT
+
+### Patches
+
+- Migrate to ok-ex-button ([ni/nimble@302334b](https://github.com/ni/nimble/commit/302334bcd302adc47b77eb5a36197f8ae67dec4e))
+
+## 0.4.3
+
+Tue, 21 Apr 2026 03:27:50 GMT
+
+### Patches
+
+- Move files into fv project folders ([ni/nimble@7a1bbdb](https://github.com/ni/nimble/commit/7a1bbdbb9ddd6e59424769d1e10490af32dfd5d6))
+
+## 0.4.2
+
+Mon, 20 Apr 2026 22:42:19 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.16.0
+
+## 0.4.1
+
+Mon, 20 Apr 2026 16:30:31 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.6.1
+- Bump @ni/spright-components to v6.15.9
+
+## 0.4.0
+
+Sat, 18 Apr 2026 13:15:22 GMT
+
+### Minor changes
+
+- feat: add OK accordion-item ([ni/nimble@0e9152f](https://github.com/ni/nimble/commit/0e9152f986e53a8fd371cdec3f805923c0e663f7))
+
+## 0.3.13
+
+Wed, 15 Apr 2026 19:52:21 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.6.0
+- Bump @ni/spright-components to v6.15.8
+
+## 0.3.12
+
+Wed, 15 Apr 2026 19:06:54 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.5.8
+- Bump @ni/spright-components to v6.15.7
+
+## 0.3.11
+
+Tue, 14 Apr 2026 20:31:57 GMT
+
+### Patches
+
+- Update fast dependency to fix view bug in Firefox ([ni/nimble@c2354bd](https://github.com/ni/nimble/commit/c2354bd5a997f90670c66a7c97d7e858607d2000))
+- Bump @ni/nimble-components to v35.5.7
+- Bump @ni/spright-components to v6.15.6
 
 ## 0.3.10